🔨 Refactor Deployment Workspace to use hooks
