+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/index.js
@@ -12,7 +12,7 @@ import { GitHub } from './items/GitHub';
 import Server from './items/Server';
 import { Live } from './items/Live';
 import More from './items/More';
-import Deployment from './items/Deployment';
+import { Deployment } from './items/Deployment';
 import ConfigurationFiles from './items/ConfigurationFiles';
 import NotOwnedSandboxInfo from './items/NotOwnedSandboxInfo';
