+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/Deployment/index.ts
@@ -1,0 +1,1 @@
+export { Deployment } from './Deployment';
